New control loop and example project showing it.
It became too painful to deal with SWT calls being made in the main
thread. They need to be done from the FX thread which means everything
needs to be wrapped with the Application launch. That breaks cross
platform compatability, but it was inevitable. At least it's localized
to startup code, which for the Workbench is manageable.

New javafx examples plugin includes JavaFXControlExample to show.